move hash fields out of Commit base and into the Crdt specific class.
