+++ src/Crdt/Commit.cs
@@ -9,16 +9,23 @@ public class Commit : CommitBase<IChange>
 {
     [JsonConstructor]
     protected Commit(Guid id, string hash, string parentHash, HybridDateTime hybridDateTime) : base(id,
-        hash,
-        parentHash,
         hybridDateTime)
     {
+        Hash = hash;
+        ParentHash = parentHash;
     }
 
     internal Commit(Guid id) : base(id)
     {
+        Hash = GenerateHash(NullParentHash);
+        ParentHash = NullParentHash;
     }
 
+    public void SetParentHash(string parentHash)
+    {
+        Hash = GenerateHash(parentHash);
+        ParentHash = parentHash;
+    }
     internal Commit() : this(Guid.NewGuid())
     {
 
@@ -26,4 +33,10 @@ public class Commit : CommitBase<IChange>
 
     [JsonIgnore]
     public List<ObjectSnapshot> Snapshots { get; init; } = [];
+
+    [JsonIgnore]
+    public string Hash { get; private set; }
+
+    [JsonIgnore]
+    public string ParentHash { get; private set; }
 }
